(Buggy) Setting up final maze generation
Maze generator should take other rooms into account when building out, and should be able to tell if it's within any of the 3 areas.
However, some rooms still aren't being built, it is prone to crashing, and when removing hallways from the "roomsStillOpen" array, it can say it's "out of bounds".
I have created and set distance checks for if the maze should create a 2 way room or a dead end, though those numbers are wide so there's typically a lot of 2 way rooms.
Will need to collaborate on a fix to get it working as intended