(no sfx?) Knockback, moved hit code from enemy to hammer
Moved the code that detects the hammer hitting the enemy from the enemy script onto the hammer. Also deactivated kinematic on the enemy rigidbodies so they can be properly knocked back.
Also, removed the sound effects of the hammer hit sounds in the process and tried to attach an SFX event object to the hammer, which doesn't seem to play when hitting anything...