Performance increase with model combining
Added script to combine the book models into one mesh during runtime. Some leftover bugs are that books may be off the shelf sometimes, usually if the shelf is rotated, or static books may not be there at all.
To switch back to the individual models, you can go into the BookShelf_v1 prefab and switch the active state of the Static Books and Combined Books.