Magic Level Up system/UI update
Magic now has defined levels that they can go between, and can be seen in game. Can only really be seen and used by the PlayerStats object and the new UI script, so some items may need to be adjusted to fit these levels. (Plan for the future is to use an enum to tell what each level is supposed to be, so books and UI only need to know the level and not the point values)