Details section & fixed held item object bug
Items now have a details section when they are picked up in the inventory. Hopefully they should be activated on hover, but they work for now when the player picks up the object, too.

Also fixed up the bug where the player may place or swap an object in their inventory and the item may disappear, now the player can place or remove anything anywhere and the held item will always be properly shown!